Update options API
Update the Options REST client against the current Gate API v4 documentation. Add request-object overloads, wire the order amend endpoint, fix timestamp/query parameter handling, refresh options response models and enums, and update changelog and examples.

+++ src/Gate.IO.Api/Options/Responses/GateOptionsOrderBook.cs
@@ -53,5 +53,5 @@ public record GateOptionsOrderBookEntry
     /// Size
     /// </summary>
     [JsonProperty("s")]
-    public decimal Size { get; set; }
+    public long Size { get; set; }
 }
